Add volume support - storing images locally in ./data
Using a local volume allows to kill your registry and restart it
without losing all your images.

This is generally what you want, and even more so if you use Docker for Mac and
have to reset it to default often because of a current bug on reclaiming disk
space.

If you don't want this behavior, disable the volume.

If you want to reclaim some disk space after a while because you don't use all
images, simply prune images under ./data, or empty the folder entirely.